Refactored name of app stubs into segregation
