+++ commercialoperator/components/organisations/models.py
@@ -35,9 +35,9 @@ from commercialoperator.components.organisations.emails import (
     send_organisation_request_email_notification,
     send_organisation_request_link_email_notification,
 )
-from commercialoperator.components.stubs.decorators import basic_exception_handler
-from commercialoperator.components.stubs.mixins import MembersPropertiesMixin
-from commercialoperator.components.stubs.utils import (
+from commercialoperator.components.segregation.decorators import basic_exception_handler
+from commercialoperator.components.segregation.mixins import MembersPropertiesMixin
+from commercialoperator.components.segregation.utils import (
     retrieve_delegate_organisation_ids,
     retrieve_email_user,
     retrieve_organisation_delegate_ids,
