+++ commercialoperator/components/proposals/serializers.py
@@ -58,11 +58,11 @@ from commercialoperator.components.proposals.serializers_event import (
 from commercialoperator.components.organisations.serializers import (
     OrganisationSerializer,
 )
-from commercialoperator.components.stubs.utils import (
+from commercialoperator.components.segregation.utils import (
     retrieve_email_user,
 )
 from commercialoperator.components.users.serializers import UserAddressSerializer
-from commercialoperator.components.stubs.serializers import (
+from commercialoperator.components.segregation.serializers import (
     SegregationBaseSerializer,
     EmailUserRoSerializer,
 )
