+++ commercialoperator/components/segregation/api.py
