+++ commercialoperator/components/segregation/classes.py
