+++ commercialoperator/components/segregation/decorators.py
