+++ commercialoperator/components/segregation/mixins.py
@@ -1,4 +1,4 @@
-from commercialoperator.components.stubs.utils import (
+from commercialoperator.components.segregation.utils import (
     retrieve_email_user,
     retrieve_group_members,
 )
