+++ commercialoperator/settings.py
@@ -77,15 +77,15 @@ INSTALLED_APPS += [
     "ledger_api_client",
     "django_cron",
     "commercialoperator",
+    "commercialoperator.components.approvals",
+    "commercialoperator.components.bookings",
+    "commercialoperator.components.compliances",
     "commercialoperator.components.main",
     "commercialoperator.components.organisations",
-    "commercialoperator.components.users",
-    "commercialoperator.components.proposals",
-    "commercialoperator.components.approvals",
-    "commercialoperator.components.compliances",
-    "commercialoperator.components.bookings",
+    "commercialoperator.components.segregation",
     "commercialoperator.components.permission",
-    "commercialoperator.components.stubs",
+    "commercialoperator.components.proposals",
+    "commercialoperator.components.users",
     "taggit",
     "rest_framework",
     "rest_framework_datatables",
